+++ src/main/java/com/matrices/Servicio_Matrices/service/model/MatrizSumaResta.java
@@ -3,9 +3,9 @@ package com.matrices.Servicio_Matrices.service.model;
 
 public class MatrizSumaResta {
 	
-	public int[][] sumarYRestarMatriz(String matrices, int tipo) {
+	public double[][] sumarYRestarMatriz(String matrices, int tipo) {
 		ProcesarMatrices procesar = new ProcesarMatrices();
-		int [][] resultado;
+		double [][] resultado;
 		int fila=0, col=0;
 		
 		String [] matriz = procesar.procesarMatriz(matrices);
@@ -13,12 +13,12 @@ public class MatrizSumaResta {
 		String [] matriz1 = procesar.generarMatriz(matriz[0]);
 		String [] matriz2 = procesar.generarMatriz(matriz[1]);
 		
-		int [][] matrizA = procesar.convertirMatriz(matriz1);
-		int [][] matrizB = procesar.convertirMatriz(matriz2);
+		double [][] matrizA = procesar.convertirMatriz(matriz1);
+		double [][] matrizB = procesar.convertirMatriz(matriz2);
 		
 		fila = matrizA.length;
 		col = matrizA[0].length;
-		resultado = new int [fila][col];
+		resultado = new double [fila][col];
 		
 		for (int i=0; i < fila; i++) {
 			for (int j = 0; j < col; j++) {
